requirements.txt
asgiref==3.7.2
dataclasses==0.6
Django==4.2
django-filter==23.5
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.4.2
iniconfig==2.0.0
isort==5.13.2
mongoengine==0.27.0
packaging==23.2
pluggy==1.3.0
pykerberos==1.2.4
pymongo==4.6.1
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
